run.py
"""Entry point for POS System."""
import os

from app import create_app

# Load environment
if os.path.exists('.env'):
    from dotenv import load_dotenv
    load_dotenv()

app = create_app(os.environ.get('FLASK_ENV', 'development'))

if __name__ == '__main__':
    debug = os.environ.get('FLASK_ENV', 'development') == 'development'
    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 5000)), debug=debug)
