requirements.txt
# POS System - Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Database
SQLAlchemy>=2.0.40
psycopg2-binary==2.9.9
# Optional - for PostgreSQL: pip install psycopg2-binary
# Optional - for MySQL: pip install PyMySQL

# Forms & Validation
email-validator==2.1.0

# Reports & Export
openpyxl==3.1.2
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
gunicorn==25.1.0

pandas==2.2.1
numpy==1.26.4
